+++ pykechain/client.py
@@ -3939,8 +3939,8 @@ class Client:
         :param description: (optional) description of the workflow to filter on
         :param scope: (optional) the scope of the workflow to filter on
         :param ref: (optional) the ref of the workflow to filter on
-        :return: a single Contexts
-        :rtype: Context
+        :return: a single Workflows
+        :rtype: Workflow
         """
         request_params = {
             "name": check_text(name, "name"),
@@ -3997,7 +3997,7 @@ class Client:
     def create_workflow(self, scope: ObjectID, **kwargs) -> Workflow:
         """Create a new Defined Workflow object in a scope.
 
-        See `Workflow.create_workflow` for available parameters.
+        See `Workflow.create` for available parameters.
 
         :return: a Workflow object
         """
@@ -4049,22 +4049,98 @@ class Client:
 
     def create_stored_file(
         self,
-        name: Text,
-        filepath: Text,
-        scope: Optional[Union[Scope, ObjectID]] = None,
-        category: Optional[StoredFileCategory] = StoredFileCategory.GLOBAL,
-        classification: Optional[StoredFileClassification] = StoredFileClassification.GLOBAL,
         **kwargs) -> StoredFile:
         """Create a new Stored File object in a scope.
 
-        See `Workflow.create_workflow` for available parameters.
+        See `StoredFile.create` for available parameters.
 
-        :return: a Workflow object
+        :return: a StoredFile object
         """
-        return StoredFile.create(client=self,
-                                 name=name,
-                                 category=category,
-                                 filepath=filepath,
-                                 classification=classification,
-                                 scope=scope,
-                                 **kwargs)
+        return StoredFile.create(client=self, **kwargs)
+
+    def stored_file(
+        self,
+        name: Optional[str] = None,
+        pk: Optional[ObjectID] = None,
+        scope: Optional[Union[Scope, ObjectID]] = None,
+        category: Optional[StoredFileCategory] = None,
+        classification: Optional[StoredFileClassification] = None,
+        description: Optional[str] = None,
+        ref: Optional[str] = None,
+        **kwargs,
+    ) -> StoredFile:
+        """
+        Retrieve a single Stored File
+
+        .. versionadded:: 4.7.0
+
+        :param pk: (optional) retrieve a single primary key (object ID)
+        :param name: (optional) name of the stored file to filter on
+        :param category: (optional) category of the stored file to search for
+        :param classification: (optional) classification of the stored file to search for
+        :param description: (optional) description of the stored file to filter on
+        :param scope: (optional) the scope of the stored file to filter on
+        :param ref: (optional) the ref of the stored file to filter on
+
+        :return: a single StoredFiles
+        :rtype: StoredFile
+        """
+        request_params = {
+            "name": check_text(name, "name"),
+            "id": check_uuid(pk),
+            "category": check_enum(category, StoredFileCategory, "category"),
+            "classification": check_enum(
+                classification, StoredFileClassification, "classification"),
+            "description": check_text(description, "description"),
+            "scope": check_base(scope, Scope, "scope"),
+            "ref": check_text(ref, "ref"),
+        }
+
+        if kwargs:
+            request_params.update(**kwargs)
+        return StoredFile.get(
+            client=self, **clean_empty_values(request_params, nones=False)
+        )
+
+    def stored_files(
+        self,
+        name: Optional[str] = None,
+        pk: Optional[ObjectID] = None,
+        scope: Optional[Union[Scope, ObjectID]] = None,
+        category: Optional[StoredFileCategory] = None,
+        classification: Optional[StoredFileClassification] = None,
+        description: Optional[str] = None,
+        ref: Optional[str] = None,
+        **kwargs,
+    ) -> List[StoredFile]:
+        """
+        Retrieve a single Stored File
+
+        .. versionadded:: 4.7.0
+
+        :param pk: (optional) retrieve a single primary key (object ID)
+        :param name: (optional) name of the stored file to filter on
+        :param category: (optional) category of the stored file to search for
+        :param classification: (optional) classification of the stored file to search for
+        :param description: (optional) description of the stored file to filter on
+        :param scope: (optional) the scope of the stored file to filter on
+        :param ref: (optional) the ref of the stored file to filter on
+
+        :return: a list of StoredFiles
+        """
+        request_params = {
+            "name": check_text(name, "name"),
+            "id": check_uuid(pk),
+            "category": check_enum(category, StoredFileCategory, "category"),
+            "classification": check_enum(
+                classification, StoredFileClassification, "classification"),
+            "description": check_text(description, "description"),
+            "scope": check_base(scope, Scope, "scope"),
+            "ref": check_text(ref, "ref"),
+        }
+
+        if kwargs:
+            request_params.update(**kwargs)
+        return StoredFile.list(
+            client=self, **clean_empty_values(request_params, nones=False)
+        )